Impl#1. Board와 Comment, User 연동 - Comment 작성시 자동으로 작성자 표시 연동 - Comment 삭제 권한 해당 유저에게만.

+++ src/main/java/boardStudy/boardStudy/service/CommentService.java
@@ -3,20 +3,23 @@ package boardStudy.boardStudy.service;
 import boardStudy.boardStudy.domain.Board;
 import boardStudy.boardStudy.domain.Comment;
 import boardStudy.boardStudy.repository.CommentRepository;
+import boardStudy.boardStudy.repository.UserRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 @Service
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final UserRepository userRepository;
 
-    public CommentService(CommentRepository commentRepository) {
+    public CommentService(CommentRepository commentRepository, UserRepository userRepository) {
         this.commentRepository = commentRepository;
+        this.userRepository = userRepository;
     }
 
-    public Comment saveComment(Comment comment, Board board) {
-        //author 부분은 나중에 처리할 것.
-        comment.setNickname("익명");
+    public Comment saveComment(Comment comment, Board board, Long userId) {
+        comment.setUserId(userId);
+        comment.setNickname(userRepository.findById(userId).get().getNickname());
         comment.setBoard(board);
         return commentRepository.save(comment);
     }
@@ -29,11 +32,20 @@ public class CommentService {
         return commentRepository.findAll();
     }
 
-    public void deleteCommentById(Long id) {
-        commentRepository.deleteById(id);
+    public void deleteCommentById(Long id, Long userId) {
+        if (commentRepository.findById(id).get().getUserId().equals(userId)) commentRepository.deleteById(id);
     }
 
     public List<Comment> getCommentsByBoardId(Long boardId){
-        return commentRepository.findByBoardId(boardId);
+        List<Comment> commentList = commentRepository.findByBoardId(boardId);
+        for (Comment comment : commentList){
+            Long userId = comment.getUserId();
+            String currentNickname = userRepository.findById(userId).get().getNickname();
+            if (!comment.getNickname().equals(currentNickname)){
+                comment.setNickname(currentNickname);
+                commentRepository.save(comment);
+            }
+        }
+        return commentList;
     }
 }
